Update from upstream repo dklollol/Olsen-Framework-Arma-3@master (#4)
* Module Additions

Added modules:
cvc_check
gas_script
radio_messages
random_scenario
remove_map
shot_count_classic

* Correcting locality issue with aCount_textX

aCount_textBLU and aCount_textRED were not declared on the clients, and so were throwing script errors

* Rewrote CoverMap Module

updated CoverMap module to include centering the map upon briefing and in game start to the center of the AO instead of on player start location - includes the ability to switch between different AOs, intent for pilots to be able to navigate to an airfield outside the AO, or for different Map AOs for different player slots/teams.

* Added Start in Vehicle Module

allows for a JIP friendly way of starting a player in a specific spot in a vehicle on mission start. Includes null/!alive vehicle checks, seat occupation, wrong define checks. Dependencies: jip module.

* Changed function call from postInit to CBALoadingscreen event

* Removed "flash" of map on game init

Changed Execution of LiveChange to waituntil visiblemap

* EnableDebugConsole Change

The current setting is 0 - Default behavior, available only in editor, changing this attribute to 1 - Available in SP and for hosts / logged in admins may fix some issues with the GM console.

* Capturezone pathing fix

* Fixed map flash at mission start

changed init to on map open

* fixed unflipping and added engine off option in ai drivers

* fixed ambiguous argument

* added wave respawn, changed folder names

* Updated FM Documetnation

* added independent and civilian to start text

* True/false 1/0 bugfix for AI attack

* Fix to eg_spectator_mode and disconnect_Control module duplicate folders

* Addition of dev branch commit.

* Tweaks to settings.sqf, and the inclusion of a new addRack functionality

* readability is important

* Added clarification to documentation comments

* removed duplicate modules